Bug#16361170 - assertion failed: join->best_read < 1.7976931348623158e+308, file sql_planner.cc
Problem:-
In Optimize_table_order::greedy_search, When 'best_read < DBL_MAX'   
means that optimizer managed to find some join order, but when 
we are not able to choose any order, it assert.

Analisys:
In Item_sum::update_used_tables(), we are setting used_tables_cache. 
But it is correct as long as aggregation is done within the same 
query block. But in case when it is refering to outer query block. 
We are setting to according to outer query tables.
That's why we can't able to choose the order in 
Optimize_table_order::best_extension_by_limited_search() and 
had assert.

Solution:-
In case, when we refering to outer_query block, we need to set it 
to OUTER_REF_TABLE_BIT.