Problem: An error message shows a variable that is not available to change.
Analysis: 

Several error messages have an explicit reference to 
@@SESSION.GTID_NEXT_LIST. 
The issue is that SESSION.GTID_NEXT_LIST is not 
available to be changed and nor it should be, since it is a
future feature to be integrated in NDB, thus not being 
available in the server mainline. The error messages are 
misleading to a user.

Fix: 
Change the error messages in order to state the only 
possible source of error. Also corrected the binlog
tests that were expecting that specific message.