Bug #13927679 	TEMPORARY ERROR 899 'ROWID ALREADY ALLOCATED' FROM NDBCLUSTER
The TUP component of the data node handles ROWID allocation, and is 
involved in query processing.

When INSERTing a row, TUP must allocate a fixed-size slot in a fragment,
denoted by a ROWID, as well as any required variable-sized space, disk space
etc.
When INSERTing, UPDATing or DELETing rows, TUP must update any ordered 
indices.

Bugs were found in the handling of errors in 2 situations :
  1.  When ordered index update fails due to the ordered index being
      unable to allocate storage space for index nodes.
  2.  When disk space allocation fails due to the Tablespace being 
      full.

In these situations, the TUP error handling did not correctly release
the fixed-size slot in the fragment.  This leak wasted space, which
could only be reclaimed by restarting the data node.

Where NoOfReplicas >= 2, there are multiple replicas of every fragment.
For INSERT, the primary replica chooses a free fixed sized slot, referenced
by a ROWID.  It then forwards the INSERT to the backup replica(s), who
must use the same ROWID.

Where the backup replica discovers that the ROWID is not available, it
will return error 899, indicating a mismatch between Primary and Backup.

The Primary and Backup roles can switch around due to node failure and 
recovery.

Where the backup fragment replica has some leaked ROWID slots, these
cause problems when the Primary replica uses them and asks the backup
to do the same.

Where the primary fragment replica has some leaked ROWID slots, these
should not cause problems as it will not attempt to use them for INSERTs.

This patch fixes the error handling for these scenarios, and enhances
an existing testcase (testBasic -n TupError) to expose the problem and 
verify the fix.

This fixes all currently known causes of error 899.