Bug #16436567:REGRESSION: CRASH AFTER JOIN::MAKE_TMP_TABLES_INFO()
Problem:
While creating temporary table, server crashes checking if
optimizer is using loose_index_scan.
     
Analysis:
In get_best_combination(), join_tab structures are created
based on the number of primary tables, semi-join nests and
temp_tables. While calculating the number of temp_tables,
optimizer estimates it to be 'one' for the query
"select ( select distinct q.a+ variance(g.a) from g q ) from g; "
as there is only distinct in it.  As a result space for only
two join_tab structures is allocated(one for the primary
table and one for the tmp table).
    
In make_tmp_tables_info(), we see that
"using_indirect_summary_function" is enabled and as a result
optimizer has to create the second tmp table. Here, its presumed
that the join_tab structure for the second tmp table is
already allocated and hence it tries to write on to this memory
while creating the intermediate temp table.
In doing so, it over-writes some of the  memory allocated for
"select" and later leads to crash.
   
Solution:
Calculate the number of outer aggregation functions while
counting field types and use the same to allocate the second
join_tab structure in get_best_combination.