Bug#16983143 - CRASH WITH STORED ROUTINE, SELECT FROM CURSOR,                UNION
Analysis:
-----------
The problem was that fetching rows from CURSOR of SELECT 
having UNION was crashing. The column meta-data stored
with cursor had a dangling pointer "orig_table". Accessing
this member was resulting in crash.

The SELECT query of CURSOR used a intermediate table for
the UNION. Then, Select_materialize::send_result_set_metadata
created a temporary table for storing the query result. The
column meta-data for the query is obtained by calling function
get_field_list. get_field_list returns "st_select_lex_unit::
item_list" for UNION, which is a list of fields which 
pointing to intermediate temporary table of union query. So, while
creating the temporary table to store query result for cursor,
the orig_table was set as intermediate temporary table of the
union. "Materialized_cursor:item_list" was also populated
with Materialized tables field meta-data. But at the end of
open cursor statement execution, intermediate temporary
table created for union is dropped. Because of this "orig_table"
became dangling.

Fix:
------
The problem is solved by using the column meta-data list
"st_select_lex_unit::types" when creating the Materialized table
of cursor instead of the column meta-data of the intermediate
table created for union query.