bug#16931177 BGC and slave binlog rotate causes slave sql_thread to stop bug#16594095 NOT VERBOSE ENOUGH ERROR REPORTING BY MTS WORKER
A slave applier thread stops with an error when it is about to rotate binlog but
finds it's already done by a thread before it in the committing group.

Notice that it's the slave applier that reacts to such situation
even though the regular use session also senses it in MBL::commit().
But unlike the slave applier the "master" user session does not propagate
any error to the top level (which is actually an issue being addressed by Bug16579083).

It appears that the reason of this trouble is actual overreaction.
About to rotate session should not raise any error when it's merely beaten to by
some other.
And that is fixed.
******
bug#16594095 NOT VERBOSE ENOUGH ERROR REPORTING BY MTS WORKER

In case a Worker fails to apply an event the Slave applier (SQL thread +
Workers) stops. An error that the Worker faced is displayed in
Show-Slave-Status. However unlike the sequential execution there's 
no event coordinates that makes the failed event identification rather difficult.

Fixed with adding a new prefix part to Last_SQL_Error message that
carries the physical (master binlog) coordinates and GTID_NEXT when one is set.
In case of simultaneous multiple Worker failures the error log must be consulted
as well.

******
merge from 5.6 repo.