Bug#17063675 USE OF UNINITIALIZED PTHREAD_GETSPECIFIC() KEY IN DEBUG BUILDS
Before this fix, code that uses DBUG_ENTER() and related macros,
when executed very early in the server initialization, typically
when using C++ constructors in static code, could end up executing
my_pthread_getspecific() with a non initialized pthread_key.

The behavior in this case is undefined, and can potentially cause a crash.

This fix is a backport from MySQL 5.7,
which uses a THR_KEY_mysys_initialized variable to keep track of 
the initialization state.

With this fix, calling DBUG_ENTER() and related macro is more robust,
and does nothing when the debug component is not available.