Bug#17041705 DO NOT BUILD PERFORMANCE SCHEMA FOR EMBEDDED
Before this fix, the performance schema storage engine was build for the
embedded library.

There is no performance instrumentation when building for embedded,
this is dead code.

The fix is to introduce a new option to the MYSQL_ADD_PLUGIN cmake macro,
named NOT_FOR_EMBEDDED.

When this flag is set,
- the plugin is not built for the embedded server,
- the plugin is conditionally included (#ifndef EMBEDDED_LIBRARY)
  in the sql/sql_builtin.cc generated file.
- the plugin is not in the list of libraries used to build libmysqld

Now, because some instrumentation can be expanded at compile time,
in particular with PSI_XXX_CALL macros when building code,
the option RECOMPILE_FOR_EMBEDDED is also added to plugins that
can contain references to the server itself (define MYSQL_SERVER).

This affects the blackhole and the federated engine.

Last, when executing tests with MTR, when using the --embedded server option,
FRM files for performance schema tables could still be present on disk,
which affects queries on the information_schema,
causing spurious ER_UNKNOWN_STORAGE_ENGINE warnings.

The root cause was that the installation process,
when using mysqld --bootstrap to install a new database,
was always creating performance schema tables (and FRM files),
because the performance schema plugin was always loaded.

mysqld --bootstrap has been relaxed a bit,
to allow for not always loading the performance schema storage engine.

The MTR script, when installing a database for an embedded test,
now installs a database without performance_schema tables.