BUG#16358989 - BIT(0): ILLEGAL, BUT PRODUCES NO ERROR OR WARNING
DESCRIPTION:

If the size for the BIT data type is specified as zero, MySQL
accepts it without any error or warning. It implicitly converts
size to one. The range of the size for BIT data type is 1 to
64. Default value for the size of BIT data type is 1.

FIX:

Changed the behavior to give an error if size for the BIT data
type is specified as zero.