Bug #12368577 MY_STRNNCOLLSP_UTF16: ASSERTION `(SLEN % 2) == 0' FAILED
Problem Description: mysql server gets assertion failure with the below set
of queries in debug build.

Analysis:
a) set @a='a'; - Here the charset value is _utf8 and the length is '1'. 
These values will be stored for the variable '@a' (by "variable" we mean 
user_var_entry object).

b) select 0 &&  @a:=_utf16'q'; - Here the charset is getting updated in 
the prepare stage hence '@a' gets new charset value i.e _utf16. In the 
execution stage the Server evaluates the expression (0 &&  @a:=_utf16'q') 
and sees '0'. Hence, the value and the length is not updated. But the 
charset for '@a' was changed to '_utf16' at preparatory stage and remains 
as it is. We can observe that change with ("SELECT @a,CHARSET(@a),LENGTH(@a);")
before this step and after this step.

c) SELECT ''=@a; - If we give this statement then it will try to compare @a 
using my_strnncollsp_utf16() function. Because the charset of '@a' is seen 
as '_utf16' so the _utf16 comparison function is chosen. Since the execution 
of the 'b' statement is not complete because of '0' in the expression the 
length of '@a' remains '1'. But per the design _utf16 should have length 
as multiples of 2. So, the query ends in debug assert.

Fix: Before fix the charset change is happening in the query preparation 
stage and also at the execution stage. 
The fix is having the charset change for user variables from 
args[0]/save_result *only* at execution phase, through 
Item_func_set_user_var::update() function. The appropriate code already 
exists in the current code base. In the uploaded patch I have initialized
the charset value for the user variables at the time of variable 
creation time. That solves the issue.