Bug#16083211 - DO NOT EXPLICITLY DISABLE PERSISTENT STATS WHEN RUNNING 	       IN READ ONLY MODE
Problem:
--------
Persistent stats were disabled when running in read only mode. This is
not necessary because we would never try to save the stats on disk when
running in read only mode (unless ANALYZE TABLE is ran) and fetching of
the stats from disk will work as it does not involve any modification of
on-disk data.

Fix:
----
Enabled Persistent stats for read only mode. This means we do not try to
update the persistent stats in read-only mode, but the stats data is used
if it exists.

Approved by Vasil. rb#2326