BUG#17154390 - CRASHING WHEN EXECUTING SELECT ON                NEW PFS TABLES WITHOUT REPLICATION
Problem:
========
There is a test in perfschema suite which tests the same.
But the issue was not caught because mtr starts the server
with server-id=1. Now the code says: 
if(server-id!=0) 
  init_slave(); 
Since server-id=1, init_slave is called and active_mi is
declared and thus dereferencing an attribute of active_mi
is fine. 

When server is started manually, server-id=0. Hence,
init_slave() is not called, so active_mi is not declared.
So, de-referencing an attribute of active_mi lead to a
server crash.

Fix:
===
Fixed the issue by checking for NULL before dereferencing.