Fix Bug#17057168 LARGE PERFORMANCE REGRESSION FOR INNODB GEOMETRY/SPATIAL INDEX LOOKUP
For the new datatype DATA_GEOMETRY, comparison of this data always
return 0. This caused it can't find the correct key range of the index
on this data type, so the query on this type column will always do full
index scan. That's why we got this performance regression.
The solution is: Since we still use still use BLOB as underlying
datatype, so we should compare geometry data following the compare BLOB
way.
In this patch, I also fixed the occasionally failure of test case
i_innodb.innodb_bug15963619. According to Krunal's suggestion, I
increased the restart wait time from 10 to 60.

rb#2890 Approved by Jimmy.