Bug#17164720 ASSERT IN READ_TOKEN , STORAGE/PERFSCHEMA/PFS_DIGEST.H
Before this fix,
  SELECT * from performance_schema.events_statements_current
could cause an assertion failure when computing the DIGEST / DIGEST TEXT columns.

The root cause is a race condition, that happen under load.
The thread computing the digest reads data from a buffer,
which is also conccurently updated when the target session is executing a statement.

The fix sanitizes data when reading from the digest storage buffer,
and when data that does not fit the expected format is found,
a NULL digest and digest text is displayed.