BUG#17160671: SLAVE_PARALLEL_WORKERS INCOMPATIBLE WITH RELAY_LOG_INFO_FILE
When setting relay-log-info-file and slave-parallel-workers > 0, then
the slave will use the value set through relay-log-info-file to create
each worker thread relay-log.info file. However, currently the slave
is concatenating the string "worker-" with the contents of
relay-log-info-file, regardless whether it is a full path or not.

This patch fixes the problem by extracting the directory part of the
value set for relay-log-info-file and uses it as a prefix for the
string "worker-". Then the part of the value that is not a directory
is used as a prefix for naming the worker relay-log.info file:

dirpart(relay-log-info-file)+"worker-"+namepart(relay-log-info-file)

If dirpart(relay-log-info-file) is empty, then the worker info files
will be created in the the datadir.