BUG#16662771: MULTIPLE RPL TESTS FAILING ON WINDOWS DUE TO CRASH AT PROTOCOL::STORE()
The problem manifests itself when a user thread is issuing a SHOW
SLAVE STATUS while at the same time the slave IO thread is being
terminated because of some error (in the case analyzed, the error was
because the IO thread had received a packet bigger than
slave_max_allowed_packet).

Looking at handle_slave_io, specifically at the part that handles
errors, one can find that the THD structure is deleted before
mi->info_thd is set to NULL. That is not an issue in itself, but the
fact that the deletion happens while it is not guarded by
mi->info_thd_lock makes it a problem. As a consequence, a race for the
THD memory area, between the user thread and the IO thread that is
shutting down and cleaning its data structures, can happen. To be
precise, if the user thread accesses THD memory through mi->info_thd
and before that the IO thread frees the memory but has not yet set the
info_thd to NULL, then the server can crash.

To fix this problem the deletion of the data structure while the IO
thread is shutting down is moved to a later stage of this process,
i.e., after setting mi->info_thd to NULL.