Bug #16579083  SESSION ATTACH ERROR WILL CAUSE INCONSISTENCY BETWEEN MASTER AND SLAVE
In group commit, the session attach error will cause the rollback
of the transaction, but the transaction is binlogged and will be
replicated to slave at that time. Which will cause inconsistency
between master and slave finally.

To fix the problem, we attach a POSIX thread to a session in
ten tries when encountering 'out of memory' error, in case
the 'out of memory' is temporary. Terminate the server after
failed to attach the POSIX thread to the session in ten tries.
Then the consistency between master and slave can be guaranteed
after mysql server restarts, as the transaction is binlogged
at that phase and the transaction is committed during recovery.
To a statement with non-transactional table, we don't guarantee
the consistency after server restarts, as it is not committed
during recovery.