Bug #17164074: TABLE_MAP EVENT'S DATA SIZE IS NOT SET WHEM READING IT FROM A PACKET.
Currently, when the table map event is built during execution, its
data size is calculated and later used when the event is written to
the binary log. The same does not happen when the event is created
from a net packet however. In fact when the event is deserialized, its
fields are set but not its size. This can in the future cause problems
if we for some reason write this event to a cache or binlog as its
size will be 0 when requested.

The size is now calculated in both situations to avoid future problems