Bug#16820156 MASTER_DELAY DOES NOT PERFORM PROPER TYPE CHECKING Bug#16960315 ERROR 1729 OVERFILLS BY THE VALUE OF MASTER_DELAY GREATER THAN 2^31
Problem:
=======
The problem happens when users try to set MASTER_DELAY using 
CHANGE MASTER TO command. There are two different issues
reported:

1) When the value to set is greater than 2^32 - 1, the server
will raise an error message stating that the user was trying
to set an value that exceeds the maximum, but the number shown
in the message will be other than the specified in the command.

2) When the value to set is greater than 2^64 - 1 and in the
following format:
"184467440XXXXXXXXXYY", where XXXXXXXXX > 737095516 and YY <= 15,
the command is accepted without presenting errors and the value
assigned is equal to the argument passed minus 2^64.

Analysis:
========
The accepted range of MASTER_DELAY is between 0 and
MASTER_DELAY_MAX (2^31 - 1).

The MASTER_DELAY_SYM is defined as ulong_num on parser.

Before assigning the MASTER_DELAY, the parser checks if the argument
informed by the user is greater than MASTER_DELAY_MAX, raising
error ER_MASTER_DELAY_VALUE_OUT_OF_RANGE if it is the case.

The error message ER_MASTER_DELAY_VALUE_OUT_OF_RANGE receives
the unaccepted value as an unsigned int value to print the message
to the user.

The parser accepts any kind of number as ulong_num and will
compare the passed argument with MASTER_DELAY_MAX correctly, but
as it is casting the value to unsigned integer when calling my_error,
only the lower 32 bits of the value are passed to the error print
function.

The second issue reported above is a bug in the function that
converts strings to unsigned long long values. When the value passed
has the pattern mentioned above, the overflow test fails and the lower
64 bits of the larger value is returned.

Fix:
===
Replaced the cast to unsigned int of the parser variable to a string
pointer of the current token evaluated by the parser.

Had also to change the parameter type in the error message to string.

A bug was opened to the second issue (Bug 16997513).

@ sql/share/errmsg-utf8.txt
  Changed the type of the invalid argument in the message to string.
  Before the fix, when the value in the argument was greater than 
  2^32 - 1, the value displayed by the message was equal the 
  argument minus 2^32 up to the unsigned log limit of the operating
  system.

@ sql/sql_yacc.yy
  Sending a string pointer of the current token evaluated instead of
  casting the out of range value to unsigned int before sending.