This is patch for bug#17170755: Wrong internal representation of trigger creation timestamp.
The bug relates to wrong storing of timestamp value in the data dictionary (.TRG file).
Test case was added to guarantee that internal storing format won't change in the future
without change in the data dictionary format.
Also minor refactoring was done - repeatable code to calculate timestamp as timeval 
was moved to dedicated method of class Trigger.