BUG#15909788-TEMPORARY FILES CREATED BY BINARY LOG CACHE ARE NOT PURGED AFTER              TRANSACTION COM
Problem:
When you modify a large amout of data inside a transaction,
for example a big LOAD DATA INFILE, and the size overcomes the binary log cache
it will create temporary files in the tmp partition. Those files don't disappear
until you close the connection, thus potentially filling up all available disk
space in MySQL configured temp directory.

Analysis:

The binlog uses an internal cache implementation (IOCACHE) to store operations
which have not yet been commited, as such, not yet written in the Binlog file.
As stated in the bug, that cache is created per connected
client, since all commit operations are intercepted by the binlog if it is
active. When binlog_cache_size overflows, a file is created in the configured
temporary directory by IOCACHE and it can grow up until
max_binlog_cache_size (4GB). It has a "greedy" optimized approach,
meaning that the create temp file is reused after a commit,
having its only file seek cursor back to 0
(mf_iocache.c::reinit_io_cache(...)), but will take the size of
the largest transaction ever made by that client. When a client leaves, and
as stated in the bug, IOCACHE is closed and the temporary file is deleted.

Fix:
The implementation was refactored in order to be intelligent and regain
resources by reducing the file size to 0 without deleting
it in the reset() method of the binlog_cache_data class, which is called
after COMMIT or ROLLBACK operations.