The assertion 'ut_ad(oldest_lsn <= cur_lsn)' can't hold, because we get the current max LSN before we get the oldest LSN from the buffer pool.
We can simply change the assertion to 'ut_ad(oldest_lsn <= log_get_lsn())'
to fix it. At the meantime, the cur_lsn is used in a subtraction, which is
'cur_lsn - oldest_lsn' to get the 'age'. If the original assertion can't hold,
the subtraction would result in a very big number. Although the unexpected
subtraction is harmless, I'd like to check the cur_lsn to get a reasonable
'age'.

rb#3013, approved by Marko