Bug#16587369: HANDLER FOR SQLEXCEPTION NOT CATCHING ERROR 1452 (FK CONSTRAINT) FOR UPDATE
Overview
--------
This is a bug regarding the handling of fatal errors. A fatal
error is an error that should not be caught by a condition
handler in a stored procedure. The issues described in the bug
report are that:

1. The same error is handled differently for different
   statements.
2. Seemingly trivial errors are treated as fatal.

It should be noted that in the context of this bug report (and
bug fix), only handler errors are relevant to consider. There are
two important issues regarding handler errors in the source code:

1. Deciding whether a handler error is to be considered fatal.
2. Ensuring that the handler error is indeed treated as fatal.

For the first issue, "is_fatal_error()" is provided by the
handler api, and returns true for handler errors that are
fatal. The second issue can be handled in various ways; the most
common is to call "print_error()" with a flag indicating that the
handler error is to be considered fatal. Inspecting how the
source code handles these two issues, in light of the problems
reported, reveals three shortcomings in the implementation:

1. Handler errors are in fact fatal by default, as defined by the
   "is_fatal_error()" function.
2. When an error is returned from the handler api, the caller
   does not always check if the error is fatal.
3. The presence of a fatal error is not always flagged when
   calling "print_error()".


Suggested fix
-------------
The scope and ambition for the fix is to treat handler errors
consistently for DML statements (update, insert, delete) by
making sure "is_fatal_error()" is called before calling
"print_error()", and to provide the "ME_FATALERROR" flag as
appopriate. This alone does not fix the bug that is reported,
since handler errors are fatal by default, hence, we also need to
extend the function "is_fatal_error()" to recognize more handler
errors as being non-fatal.

An attempt was made to consider handler errors non-fatal by
default, but this made a large number of tests fail in not very
obvious ways, so this solution was not pursued further. Instead,
a switch for catching non-fatal errors was added to
"is_fatal_error()", currently catching:

- HA_ERR_NO_REFERENCED_ROW: The error originally addressed by the
  bug report considered here.
- HA_ERR_ROW_IS_REFERENCED: Another FK constraint violation
  error, an obvious extension of the bug we are addressing.
- HA_ERR_LOCK_WAIT_TIMEOUT: See below.

Changing the implementation of "is_fatal_error()" has a potential
impact on the non-DML source code, which also calls this
function; however, no damage has been observed in the testing
that has been done. On the other hand, tests failed for other
reasons:

1. Some tests expected a fatal outcome in situations where the
   error is now non-fatal (e.g. for foreign key constraint
   violations). The fix for this was to modify the test.
2. Some tests expected a non-fatal outcome where the error is now
   fatal due to the changed implementation of DML statements
   where we ensure that errors are treated the way they ought
   to. The fix for this was to extend "is_fatal_error()" to
   consider the error in question as non fatal (the error here
   was lock wait timeout).

Re-considering which handler errors should be considered fatal is
still a relevant task, but is considered beyond the scope of this
bug fix. Additionally, such a task might be the responsibility of
the storage engine teams.