Bug#16896810 INPLACE ALTER: FAILING ASSERTION: !PREBUILT->TRX->CHECK_FOREIGNS
When there are multiple indexes on the same column, and all of them
are being dropped while a foreign key constraint needs an index,
InnoDB would not block the DROP INDEX operation. (InnoDB foreign key
constraint checks are only performed by index lookup, never by a table
scan.)

ha_innobase::prepare_inplace_alter_table(): Flag the indexes to be
dropped, before invoking innobase_check_foreign_key_index(), so that
to-be-dropped indexes will be skipped when determining if there are
replacements for all to-be-dropped indexes.

innobase_check_foreign_key_index(): Remove the assertion
ut_ad(!index->to_be_dropped). The index should be flagged as
to-be-dropped except when dropping the primary key.

rb#3020 approved by Jimmy Yang