Bug#11748775: OPTIMIZER USES WRONG INDEX FOR THE RIGHT TABLE IN LEFT JOIN
The range optimizer used the wrong prerequisite for concluding that a
table is the inner table of an outer join. It looked at
TABLE::maybe_null which, if true, means that a table is the inner
table of an outer join. However, this flag is not unset if the outer
join is rewritten to an inner join. This caused the range optimizer to
bail out on such indexes. The ref optimizer, on the other hand, used
the correct information. The effect is that the optimizer gets a cost
estimate from the ref optimizer only, and it is so far off that the
wrong index gets chosen.

The fix is to make the range optimizer correctly recognize what is
actually an inner table of an outer join. This causes it to produce a
range access plan, but - more importantly - a cost. Thanks to the
heuristic that range statistics take precedence if the ref estimate
seems too optimistic, the index gets rejected, and the obviously
better index is chosen.

Fixing the bug has the additional benefit that a lot of impossible
where conditions are now discovered that previously were not, more
precisely when there is an outer join with an IS NULL predicate on a
non-nullable column of the inner table, and the outer join is
rewritten to inner join.

A lot of test cases change. Some change because of the newly
discovered impossible where clause mentioned above, while some change
because we can now do a range scan where previously the optimizer
thought a table scan was the only option.